requirements.txt
azure-functions
requests
azure-cosmos
openai
azure-identity